payloads: Updated payloads to unsigned long return value
Payloads launched via "go" should actually be have the following
prototype [1]:

    unsigned long main(int argc, char *argv[])

[1]: https://source.denx.de/u-boot/u-boot/-/blob/v2021.07/cmd/boot.c#L18

+++ payloads/src/read_memory.c
@@ -2,7 +2,7 @@
 #include "u-boot.h"
 #include "str2uint.h"
 
-int main(int argc, char *argv[])
+unsigned long main(int argc, char *argv[])
 {
     int status;
     jt_funcs_t *jt;
